# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "Body_ingest_file_api_rag_ingest_file_post": {
        "properties": {
          "area": {
            "default": "general",
            "title": "Area",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_ingest_file_api_rag_ingest_file_post",
        "type": "object"
      },
      "Body_ingest_text_api_rag_ingest_text_post": {
        "properties": {
          "area": {
            "default": "general",
            "title": "Area",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "source": {
            "default": "manual",
            "title": "Source",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "Body_ingest_text_api_rag_ingest_text_post",
        "type": "object"
      },
      "Body_ingest_url_api_rag_ingest_url_post": {
        "properties": {
          "area": {
            "default": "general",
            "title": "Area",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "Body_ingest_url_api_rag_ingest_url_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RagQuery": {
        "properties": {
          "area": {
            "default": "general",
            "title": "Area",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "top_k": {
            "default": 3,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "RagQuery",
        "type": "object"
      },
      "RagResult": {
        "properties": {
          "area": {
            "title": "Area",
            "type": "string"
          },
          "chunk": {
            "title": "Chunk",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "chunk",
          "score",
          "source",
          "area"
        ],
        "title": "RagResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "StudioCookie": {
        "in": "cookie",
        "name": "boerdi_studio_auth",
        "type": "apiKey"
      },
      "StudioKey": {
        "in": "header",
        "name": "X-Studio-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "boerdi-chat backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/rag/area/{area}": {
      "delete": {
        "description": "Delete all chunks in a knowledge area, and its document rows.\n\nALT swept the sqlite-vec index here too, because embeddings lived in a\nseparate table whose rows could outlive their chunk. pgvector keeps the\nembedding as a column on ``rag_chunks``, so the vector dies with the chunk\nand there is nothing left to sweep (see services/rag/admin.py).",
        "operationId": "delete_rag_area_api_rag_area__area__delete",
        "parameters": [
          {
            "in": "path",
            "name": "area",
            "required": true,
            "schema": {
              "title": "Area",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Delete Rag Area",
        "tags": [
          "rag"
        ]
      },
      "get": {
        "description": "List documents in a knowledge area.\n\nDocuments are grouped by the compound key ``(title, source)`` so that\ne.g. two uploads with the same filename from different folders, or two\nmanual entries with the same title, remain distinguishable.",
        "operationId": "get_rag_area_api_rag_area__area__get",
        "parameters": [
          {
            "in": "path",
            "name": "area",
            "required": true,
            "schema": {
              "title": "Area",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Rag Area",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/area/{area}/doc": {
      "delete": {
        "description": "Delete a single document (all its chunks) from a knowledge area.\n\nQuery params ``title`` and ``source`` together identify the document\n— they must match the values returned by ``GET /area/{area}``. Empty\nstrings are valid (the match is exact).",
        "operationId": "delete_rag_doc_api_rag_area__area__doc_delete",
        "parameters": [
          {
            "in": "path",
            "name": "area",
            "required": true,
            "schema": {
              "title": "Area",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "default": "",
              "title": "Title",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "",
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Delete Rag Doc",
        "tags": [
          "rag"
        ]
      },
      "get": {
        "description": "Return all chunks of a single document, ordered by chunk_index.\n\nIdentified by ``(area, title, source)`` exactly like the delete\nendpoint. Use when the Studio wants to preview the full content\nof a RAG document instead of just the 200-char preview.\n\nReturns ``{title, source, area, chunks: [{index, content, created_at}]}``.",
        "operationId": "get_rag_doc_api_rag_area__area__doc_get",
        "parameters": [
          {
            "in": "path",
            "name": "area",
            "required": true,
            "schema": {
              "title": "Area",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "default": "",
              "title": "Title",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "",
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Get Rag Doc",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/areas": {
      "get": {
        "description": "List all knowledge areas with chunk counts.",
        "operationId": "list_rag_areas_api_rag_areas_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "List Rag Areas",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/embed": {
      "post": {
        "description": "Generate embeddings for all chunks that don't have one yet.\n\nA backfill for chunks that reached the DB without a vector (restored dump,\nmanual insert) — the ingest endpoints always embed inline. ALT additionally\nran this at startup after its seed import; NEU has no RAG seed importer, so\nthere is no such caller.\n\nReturns the number of chunks that were embedded.",
        "operationId": "rag_embed_api_rag_embed_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rag Embed Api Rag Embed Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Rag Embed",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/ingest/file": {
      "post": {
        "description": "Upload and ingest a document (PDF, DOCX, PPTX, etc.) via markitdown.",
        "operationId": "ingest_file_api_rag_ingest_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_file_api_rag_ingest_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Ingest File",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/ingest/text": {
      "post": {
        "description": "Ingest raw markdown/text into a knowledge area.",
        "operationId": "ingest_text_api_rag_ingest_text_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_text_api_rag_ingest_text_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Ingest Text",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/ingest/url": {
      "post": {
        "description": "Ingest a web page into a knowledge area via markitdown.",
        "operationId": "ingest_url_api_rag_ingest_url_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_url_api_rag_ingest_url_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Ingest Url",
        "tags": [
          "rag"
        ]
      }
    },
    "/api/rag/query": {
      "post": {
        "description": "Query the RAG knowledge base.",
        "operationId": "rag_query_api_rag_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RagQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RagResult"
                  },
                  "title": "Response Rag Query Api Rag Query Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StudioKey": []
          }
        ],
        "summary": "Rag Query",
        "tags": [
          "rag"
        ]
      }
    },
...
  }
}